+++ image/resources/knfsd-metrics-agent/internal/connections/factory.go
@@ -52,6 +52,6 @@ func createMetricsReceiver(
 		&cfg.ControllerConfig,
 		set,
 		con,
-		scraperhelper.AddScraper(metadata.Type, s),
+		scraperhelper.AddMetricsScraper(metadata.Type, s),
 	)
 }